+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityDialogs.java
@@ -497,7 +497,7 @@ public final class EntityDialogs {
 																							boolean singleSelection) {
 			EntityTablePanel tablePanel = new EntityTablePanel(tableModel);
 			tablePanel.initialize();
-			tablePanel.table().addDoubleClickListener(mouseEvent -> {
+			tablePanel.table().doubleClickObserver().addListener(() -> {
 				if (!tableModel.selectionModel().isSelectionEmpty()) {
 					okControl.actionPerformed(null);
 				}