+++ changelog.md
@@ -28,6 +28,7 @@ Codion Change Log
 - TemporalField, add/removeListener methods replaced with methods that return the underlying event observer.
 - CalendarPanel, add/removeListener methods replaced with methods that return the underlying event observer.
 - ColumnConditionPanel, add/removeListener methods replaced with methods that return the underlying event observer.
+- FilteredTable, add/removeListener methods replaced with methods that return the underlying event observer.
 
 ## 0.17.34
 ### is.codion.common.core